🔴 AlertViewer.js 🟢 AlertViewer.jsx 🛠️ AlertViewer.module.css -> Commented outline style
