+++ src/app/components/AlertViewer/AlertViewer.module.css
@@ -1,3 +1,3 @@
 .AlertViewer {
-  outline: 1px solid pink;
+  /* outline: 1px solid pink; */
 }
